python_scripts/automations/save-logs/requirements.txt
robonomics-interface==0.8.0
pinatapy-vourhey==0.1.3
PyYAML==5.4.1